config/requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
Django==5.1.7
djangorestframework==3.15.2
iniconfig==2.0.0
msgpack==1.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-asyncio==0.25.3
python-dotenv==1.0.1
redis==5.2.1
sqlparse==0.5.3
websockets==15.0.1
